requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
requests==2.31.0
proxmoxer==2.0.1
docker==7.0.0
python-nmap==0.7.1
gunicorn==21.2.0
psutil==5.9.6
